+++ bin/application/Sample.fxml
@@ -1,0 +1,104 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.RowConstraints?>
+<?import javafx.scene.layout.VBox?>
+
+<BorderPane prefHeight="720.0" prefWidth="1280.0" style="-fx-background-color: blue;" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
+   <top>
+      <HBox prefHeight="100.0" prefWidth="200.0" style="-fx-border-color: black;" BorderPane.alignment="CENTER">
+         <children>
+            <ImageView fitHeight="138.0" fitWidth="143.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../img/bestbuy.png" />
+               </image>
+            </ImageView>
+            <BorderPane prefHeight="99.0" prefWidth="459.0">
+               <opaqueInsets>
+                  <Insets left="20.0" />
+               </opaqueInsets>
+               <right>
+                  <TextField prefHeight="27.0" prefWidth="419.0" promptText="Search Best Buy" BorderPane.alignment="CENTER" />
+               </right>
+            </BorderPane>
+            <BorderPane prefHeight="99.0" prefWidth="298.0">
+               <center>
+                  <Button mnemonicParsing="false" style="-fx-background-color: white;" text="Login" BorderPane.alignment="CENTER" />
+               </center>
+               <right>
+                  <Button mnemonicParsing="false" style="-fx-background-color: yellow;" text="Cart" BorderPane.alignment="CENTER" />
+               </right>
+            </BorderPane>
+         </children>
+      </HBox>
+   </top>
+   <left>
+      <VBox prefHeight="303.0" prefWidth="140.0" style="-fx-border-color: black;" BorderPane.alignment="CENTER">
+         <children>
+            <Label style="-fx-background-color: white;" text="Filter:" />
+            <Label style="-fx-background-color: yellow;" text="Price:" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="High to Low" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="Low to High" />
+            <Label style="-fx-background-color: yellow;" text="Brand:" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="Nvidia" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="EVGA" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="AMD" />
+            <Label style="-fx-background-color: yellow;" text="Colour:" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="Black" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="White" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="Silver" />
+            <CheckBox mnemonicParsing="false" style="-fx-background-color: white;" text="Red" />
+         </children>
+      </VBox>
+   </left>
+   <center>
+      <GridPane BorderPane.alignment="CENTER">
+         <columnConstraints>
+            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+         </columnConstraints>
+         <rowConstraints>
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+         </rowConstraints>
+         <children>
+            <ImageView fitHeight="172.0" fitWidth="755.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../img/geforce-rtx-3080-shop-462-l@2x.png" />
+               </image>
+            </ImageView>
+            <ImageView fitHeight="181.0" fitWidth="910.0" pickOnBounds="true" preserveRatio="true" GridPane.rowIndex="1">
+               <image>
+                  <Image url="@../img/geforce-rtx-3070-shop-630-d@2x.png" />
+               </image>
+            </ImageView>
+            <TextArea editable="false" prefHeight="58.0" prefWidth="217.0" text="Nvidia RTX 3080 $699.99" GridPane.columnIndex="1" />
+            <Button mnemonicParsing="false" style="-fx-background-color: red;" text="Out of Stock" GridPane.columnIndex="1" />
+            <TextArea editable="false" prefHeight="62.0" prefWidth="217.0" text="Nvidia RTX 3070 $599.99" GridPane.columnIndex="1" GridPane.rowIndex="1" />
+            <Button mnemonicParsing="false" style="-fx-background-color: yellow;" text="Add to Cart" GridPane.columnIndex="1" GridPane.rowIndex="1" />
+            <ImageView fitHeight="211.0" fitWidth="1267.0" pickOnBounds="true" preserveRatio="true" GridPane.rowIndex="2">
+               <image>
+                  <Image url="@../img/geforce-rtx-2080-ti-web-tech-shot-630-u@2x.png" />
+               </image>
+            </ImageView>
+            <TextArea editable="false" prefHeight="200.0" prefWidth="200.0" text="Nvidia RTX 2080 Ti $1200.00" GridPane.columnIndex="1" GridPane.rowIndex="2" />
+            <Button mnemonicParsing="false" style="-fx-background-color: yellow;" text="Add to Cart" GridPane.columnIndex="1" GridPane.rowIndex="2" />
+         </children>
+      </GridPane>
+   </center>
+   <right>
+      <VBox prefHeight="303.0" prefWidth="31.0" BorderPane.alignment="CENTER" />
+   </right>
+</BorderPane>
