+++ src/application/SampleController.java
@@ -1,0 +1,5 @@
+package application;
+
+public class SampleController {
+	
+}
